+++ src/delfino/constants.py
@@ -2,7 +2,7 @@ from enum import Enum
 
 ENTRY_POINT: str = "delfino"
 PYPROJECT_TOML_FILENAME = "pyproject.toml"
-COMMANDS_DIRECTORY_NAME = "commands"
+COMMANDS_DIRECTORY_NAME = "commands"  # deprecated, remove once core package is created
 
 
 class PackageManager(Enum):