+++ tests/integration/test_command_registry.py
@@ -1,24 +1,31 @@
+import logging
+
 import pytest
 
-from delfino.main import _CommandRegistry
-from delfino.plugin import discover_packages
+from delfino.click_utils.command import CommandRegistry
 
 
+@pytest.fixture(scope="module")
+def command_packages():
+    return list(CommandRegistry._discover_command_packages())
+
+
+@pytest.mark.usefixtures("install_fake_plugins")
 class TestPlugin:
     @staticmethod
-    def should_register_non_duplicated_plugin_commands(install_fake_plugins):
-        del install_fake_plugins
-        registry = _CommandRegistry(disabled_command={"fake-plugin-a": set(["typecheck"])})
-        packages = list(discover_packages())
-        for package in packages:
-            registry.register_command_package(package)
-        assert len(registry.as_dict()) == 3
+    def should_register_non_duplicated_plugin_commands(command_packages):
+        registry = CommandRegistry(
+            disabled_commands={"fake-plugin-a": set(["typecheck"])},
+            command_packages=command_packages,  # don't load project packages
+        )
+        assert registry.visible_commands == ["format", "lint", "typecheck"]
 
     @staticmethod
-    def should_raise_error_with_duplicated_plugin_commands(install_fake_plugins):
-        del install_fake_plugins
-        registry = _CommandRegistry(disabled_command={})
-        packages = list(discover_packages())
-        with pytest.raises(RuntimeError):
-            for package in packages:
-                registry.register_command_package(package)
+    def should_raise_error_with_duplicated_plugin_commands(caplog, command_packages):
+        caplog.set_level(logging.DEBUG)
+        CommandRegistry(disabled_commands={}, command_packages=command_packages)
+        log_msg = (
+            "Using command 'typecheck' from plugin 'fake-plugin-b'. Previously registered "
+            "by 'fake-plugin-a' plugin, which has lower priority."
+        )
+        assert log_msg in caplog.text